requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
requests==2.32.3
tomli==2.1.0
urllib3==2.2.3
